requirements.txt
Django==4.1.2
django-admin-interface==0.22.1
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-redis==5.2.0
Pillow==9.2.0
redis==4.3.4
requests==2.28.1
